feat: adjust existing rest_api to serve and interact with frontend
